ai-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
httpx==0.27.0
faster_whisper==1.0.3
openai-whisper>=20231117
torch>=2.0
transformers>=4.30
langchain==1.1.3
langchain-openai>=0.1.14
openai>=1.0.0
python-dotenv
tiktoken
transformers
torchaudio
pandas
scikit-learn
spacy
python-multipart